Fix Phase 3: Type conversion errors (decimal↔int) - 48 errors fixed (666 → 618)

+++ src/UnifiedOrchestrator/Promotion/PromotionService.cs
@@ -473,13 +473,13 @@ internal class PromotionService : IPromotionService
                 cancellationToken).ConfigureAwait(false);
 
             // Use REAL metrics from shadow test instead of comparing model metadata
-            decision.SharpeImprovement = (double)(shadowTestReport.ChallengerSharpe - shadowTestReport.ChampionSharpe);
-            decision.SortinoImprovement = (double)(shadowTestReport.ChallengerSortino - shadowTestReport.ChampionSortino);
-            decision.CVaRImprovement = (double)(shadowTestReport.ChallengerCVaR - shadowTestReport.ChampionCVaR);
-            decision.DrawdownImprovement = (double)(shadowTestReport.ChallengerMaxDrawdown - shadowTestReport.ChampionMaxDrawdown);
+            decision.SharpeImprovement = shadowTestReport.ChallengerSharpe - shadowTestReport.ChampionSharpe;
+            decision.SortinoImprovement = shadowTestReport.ChallengerSortino - shadowTestReport.ChampionSortino;
+            decision.CVaRImprovement = shadowTestReport.ChallengerCVaR - shadowTestReport.ChampionCVaR;
+            decision.DrawdownImprovement = shadowTestReport.ChallengerMaxDrawdown - shadowTestReport.ChampionMaxDrawdown;
 
             // Use REAL statistical test results
-            decision.PValue = (double)shadowTestReport.PValue;
+            decision.PValue = shadowTestReport.PValue;
             decision.StatisticallySignificant = shadowTestReport.StatisticallySignificant;
             decision.ConfidenceInterval = 0.95; // Fixed confidence interval
 
@@ -522,11 +522,11 @@ internal class PromotionService : IPromotionService
             _logger.LogError(ex, "Shadow test failed for challenger {ChallengerVersionId}", challenger.VersionId);
             
             // Fall back to metadata comparison if shadow test fails
-            decision.SharpeImprovement = (double)(challenger.Sharpe - champion.Sharpe);
-            decision.SortinoImprovement = (double)(challenger.Sortino - champion.Sortino);
-            decision.CVaRImprovement = (double)(challenger.CVaR - champion.CVaR);
-            decision.DrawdownImprovement = (double)(challenger.MaxDrawdown - champion.MaxDrawdown);
-            decision.PValue = 0.999; // Conservative fallback - assume not significant
+            decision.SharpeImprovement = challenger.Sharpe - champion.Sharpe;
+            decision.SortinoImprovement = challenger.Sortino - champion.Sortino;
+            decision.CVaRImprovement = challenger.CVaR - champion.CVaR;
+            decision.DrawdownImprovement = challenger.MaxDrawdown - champion.MaxDrawdown;
+            decision.PValue = 0.999m; // Conservative fallback - assume not significant
             decision.StatisticallySignificant = false;
             decision.ValidationErrors.Add($"Shadow test failed: {ex.Message} - using conservative metrics");
